Se especifican mejor las variables y se corrigen mas errores

+++ ProyectoFinalAppi/ADO .NET/Handler/VentaHandler.cs
@@ -14,7 +14,7 @@ namespace ProyectoFinalAppi.ADO_.NET
         //Funciones.
 
         //Eliminar Venta.
-        public static bool EliminarVenta(int id)
+        public static bool EliminarVenta(long id)
         {
             //Varibale.
             bool ventaEliminada = false;
